Add workaround for XML editor when using tabs
The XML editor has two properties, `indentationChar` and
`indentationSize`. `indentationChar` is which character to use for
indentation (as expected) but `indentationSize` is how many of the
*`indentationChar` character* to use, now *how many columns one level of
indentation should be.*

As such, If your `.editorconfig` file has `indent_style = tab` and
`indent_size = 2` (something like below for example) then the XML editor
will use *2 tabs* for one level of indentation.

    [*]
    indent_size = 2

    [*.xml]
    indent_style = tab

Each of those 2 tabs will be 2 columns wide so the individual tabs are
being *displayed* correctly, just not *placed* correctly.

This change makes it so that the XML editor's `indentationSize` set to
the `.editorconfig` file's `indent_size`'s value *only* when the
`indent_style` is `space`. When it is `tab` the XML editor's
`indentationSize` is cleared.

- See: #41